chore(protocol): move error type to protocol module
