+++ broker-core/src/test/java/io/zeebe/broker/incident/ExpressionIncidentTest.java
@@ -31,8 +31,8 @@ import io.zeebe.exporter.record.value.IncidentRecordValue;
 import io.zeebe.exporter.record.value.WorkflowInstanceRecordValue;
 import io.zeebe.model.bpmn.Bpmn;
 import io.zeebe.protocol.BpmnElementType;
+import io.zeebe.protocol.ErrorType;
 import io.zeebe.protocol.clientapi.RecordType;
-import io.zeebe.protocol.impl.record.value.incident.ErrorType;
 import io.zeebe.protocol.intent.IncidentIntent;
 import io.zeebe.protocol.intent.WorkflowInstanceIntent;
 import io.zeebe.test.broker.protocol.clientapi.ClientApiRule;
