+++ protocol/src/main/java/io/zeebe/protocol/ErrorType.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.zeebe.protocol.impl.record.value.incident;
+package io.zeebe.protocol;
 
 public enum ErrorType {
   UNKNOWN,
